Documentation for TensorQEC.

TensorQEC.Bimatrix

TensorQEC.Code832

TensorQEC.PauliString

TensorQEC.ShorCode

TensorQEC.SteaneCode

TensorQEC.SurfaceCode

TensorQEC.ToricCode

TensorQEC.clifford_network

TensorQEC.correction_pauli_string

TensorQEC.encode_stabilizers

TensorQEC.inference

TensorQEC.measure_syndrome!

TensorQEC.pauli_string_map_iter

TensorQEC.place_qubits

TensorQEC.stabilizers

TensorQEC.syndrome_inference

TensorQEC.transformed_sydrome_dict

TensorQEC.Bimatrix — Type

Bimatrix

Since the encding process may alter the generators of stabilizer group, we introduce the Bimatrix structure to store the information of encoding process. The Bimatrix structure contains the following fields

'matrix': The bimatrix representation of the stabilizers.

'Q': The matrix records the Gaussian elimination process, whcih is used to recover the original stabilizers.

'ordering': The ordering of qubits.

'xcodenum': The number of X stabilizers.

TensorQEC.Code832 — Type

Code832

Construct a [[8,3,2]] CSS code instance.

TensorQEC.PauliString — Type

PauliString{N} <: CompositeBlock{2}

A Pauli string is a tensor product of Pauli gates, e.g. XYZ. The matrix representation of a Pauli string is evaluated as

\[A = \bigotimes_{i=1}^N \sigma_{ids[N-i+1]}\]

where ids is the array of integers representing the Pauli gates. Note the order of ids is following the little-endian convention, i.e. the first qubit is the least significant qubit. For example, the Pauli string XYZ has matrix representation Z ⊗ Y ⊗ X.

Fields

ids::NTuple{N, Int}: the array of integers (1-4) representing the Pauli gates.

1: I ($σ_0$)

2: X ($σ_1$)

3: Y ($σ_2$)

4: Z ($σ_3$)

TensorQEC.ShorCode — Type

ShorCode

Construct a Shor code instance.

TensorQEC.SteaneCode — Type

SteaneCode

Construct a Steane code instance.

TensorQEC.SurfaceCode — Type

SurfaceCode(m::Int, n::Int)

Construct a surface code with m rows and n columns.

TensorQEC.ToricCode — Type

ToricCode(m::Int, n::Int)

Construct a Toric code with m rows and n columns.

TensorQEC.clifford_network — Method

clifford_network(qc::ChainBlock)

Generate a Clifford network from a quantum circuit.

TensorQEC.correction_pauli_string — Method

correction_pauli_string(qubit_num::Int, syn::Dict{Int, Bool}, prob::Dict{Int, Vector{Float64}})

Generate the error Pauli string in the coding space. To correct the error, we still need to transform it to the physical space.

Arguments

qubit_num: The number of qubits.

syn: The syndrome dictionary.

prob: The inferred error probability of each physical qubit in coding space.

Returns

ps: The error Pauli string in the coding space.

TensorQEC.encode_stabilizers — Method

encode_stabilizers(stabilizers::AbstractVector{PauliString{N}}) where N

Generate the encoding circuit for the given stabilizers.

Arguments

stabilizers: The vector of pauli strings, composing the generator of stabilizer group.

Returns

qc: The encoding circuit.

data_qubits: The indices of data qubits.

bimat: The structure storing the encoding information.

TensorQEC.inference — Method

inference(measure_outcome::Vector{Int}, code::Bimatrix, qc::ChainBlock, p::Vector{Vector{Float64}})

Infer the error probability of each qubit from the measurement outcome of the stabilizers.

Arguments

measure_outcome: The measurement outcome of the stabilizers, which is either 1 or -1.

code: The structure storing the encoding information.

qc: The encoding circuit.

p: The prior error probability of each physical qubit.

Returns

ps_ec_phy: The error Pauli string for error correction.

TensorQEC.measure_syndrome! — Method

measure_syndrome!(reg::AbstractRegister, stabilizers::AbstractVector{PauliString{N}}) where N

Measure the given stabilizers.

Arguments

reg: The quantum register.

stabilizers: The vector of pauli strings, composing the generator of stabilizer group.

Returns

measure_outcome: The measurement outcome of the stabilizers, which is either 1 or -1.

TensorQEC.pauli_string_map_iter — Method

pauli_string_map_iter(ps::PauliString{N}, qc::ChainBlock) where N

Map the Pauli string ps by a quantum circuit qc. Return the mapped Pauli string.

TensorQEC.place_qubits — Method

place_qubits(reg0::AbstractRegister, data_qubits::Vector{Int}, num_qubits::Int)

Place the data qubits to the specified position. The other qubits are filled with zero state.

Arguments

reg0: The data register.

data_qubits: The indices of data qubits.

num_qubits: The total number of qubits.

Returns

reg: The register with data qubits placed at the specified position.

TensorQEC.stabilizers — Method

stabilizers(tc::ToricCode)
stabilizers(sc::SurfaceCode)
stabilizers(shor::ShorCode)
stabilizers(steane::SteaneCode)
stabilizers(code832::Code832)

Get the stabilizers of the code instances.

TensorQEC.syndrome_inference — Method

syndrome_inference(cl::CliffordNetwork{T}, syn::Dict{Int,Bool}, p::Vector{Vector{Float64}}) where T

Infer the error probability of each qubit from the measurement outcome of the stabilizers.

Arguments

cl: The Clifford network.

syn: The syndrome dictionary.

p: The prior error probability of each physical qubit.

Returns

pinf: The inferred error probability of each physical qubit in coding space. For errored stabilizers, there are two possibilities: X or Y error. For unerrored stabilizers, there are also two possibilities: no error or Z error. For unmeasured qubits, there are four possibilities: no error, X error, Y error, Z error. Therefore the length of each vector in pinf may be 2 or 4.

TensorQEC.transformed_sydrome_dict — Method

transformed_sydrome_dict(measure_outcome::Vector{Int}, code::Bimatrix)

Generate the syndrome dictionary on the transformed stabilizers from the measurement outcome.

Arguments

measure_outcome: The measurement outcome of the stabilizers, which is either 1 or -1.

code: The structure storing the encoding information.

Returns

syn_dict: The syndrome dictionary on the transformed stabilizers. 1 is transformed to 0, -1 is transformed to 1.